move dockertest package
